Refactor: move page components to pages directory
