+++ frontend/src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import {BrowserRouter, Routes, Route} from 'react-router-dom'
-import Landing from './components/Landing';
-import SignIn from './components/SignIn';
-import Finder from './components/Finder';
-import SignUp from './components/SignUp';
-import Guest from './components/Guest';
+import Landing from './pages/Landing';
+import SignIn from './pages/SignIn';
+import Finder from './pages/Finder';
+import SignUp from './pages/SignUp';
+import Guest from './pages/Guest';
 
 function App() {
   return (<BrowserRouter>
